% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/content.r
\name{debate.content.extract.dictionary}
\alias{debate.content.extract.dictionary}
\title{Count occurences of terms defined within a dictionary}
\usage{
debate.content.extract.dictionary(txt, dictionary,
  escape = "[!\\"#$\%&()*+,./:;<=>?^_|]", lowercase = TRUE)
}
\arguments{
\item{txt}{the text in which to count occurences}

\item{dictionary}{the dictionary defining terms to count. See \code{\link{debate.example.dic}} for a example of the dictionary data structure expected by this function.}

\item{escape}{list of characters that have to be removed from the \code{text} parameter before analysis.}

\item{lowercase}{if \code{TRUE} the \code{txt} parameter will be lowercased before analysis.}
}
\description{
This function extracts from a text, represented as a character string, all occurences of terms defined in a dictionary.
}
\examples{
path.to.pdf <- paste0(find.package("debate"), "/doc/BOACG_Tome_21_article_49.pdf")
txt <- debate.pdftotext(path.to.pdf)
txt <- debate.txt.clean(txt)
dic <- debate.example.dic()
debate.content.extract.dictionary(txt, dic)
}

