% Generated by roxygen2 (4.0.1): do not edit by hand
\name{debate.example.dic}
\alias{debate.example.dic}
\title{An example of data structure used in this package for storing a dictionary}
\usage{
debate.example.dic()
}
\description{
This function returns a basic example of dictionary data structure expected to be processed by the \code{\link{debate.content.extract.dictionary}} function FIXME.
}
\examples{
dic <- debate.example.dic()
dic
}

