% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/example.r
\name{debate.example.mp}
\alias{debate.example.mp}
\title{An example of data structure used in this package for storing the members of parliament list}
\usage{
debate.example.mp()
}
\description{
This function returns a basic example of list of members of parliament list data structure expected to be processed by the \code{\link{debate.content.extract.mps}}.
The member of parliament list is a list whose each element represents a member of parliament. Each member of parliament is also represented as a list. This list have to contain three items:
1. \code{name}: the name of the member of parliament
2. \code{pattern.start}: a character string which will be used to detect a start of speech.
3. \code{pattern.stop}: a character string which will be used to detect a end of speech.
}
\examples{
mp <- debate.example.mp()
mp
}

