% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/debate.r
\docType{package}
\name{debate}
\alias{debate}
\alias{debate-package}
\title{Measuring content and relational of (political) debates}
\description{
Several dimensions are in play during a debate. This is especially the case for political debates. This package provides tools for measuring both the content and the relational dynamics of debates. For measuring content we suggest to use a thematic coding based on a dictionary. From this dictionary, the list of the actors of the debate, and the memorials in PDF files, the package can help you to count occurences of dictionnary items in each actor speechs. The relational is measured by count approval, disapproval and neutral references between actors. Using the spnet package as a back-end for the visualization, the user can display both the content and the relational aspects of the debate on the assembly map. See (ARTICLE) for more information about this methodology.
}
\references{
Deville M. and Rousseaux E. (2013). TITLE. JOURNAL. VOLUME(NUMBER), PAGEFROM-PAGETO.
}
\keyword{debate}
\keyword{map}
\keyword{mining}
\keyword{networks}
\keyword{package}
\keyword{spatial}
\keyword{text}

