% Generated by roxygen2 (4.0.1): do not edit by hand
\name{debate.relational.extract}
\alias{debate.relational.extract}
\title{Extract the relational component of a debate as suggested by REFERENCE}
\usage{
debate.relational.extract(txt)
}
\arguments{
\item{txt}{the text to be processed}
}
\description{
This function count for each member of parliament the number of (1) approval, (2) disapproval and (3) neutral references he/she has with other members of parliament during a debate. See REFERENCE for more information about this methodology.
}
\examples{
debate.relational.extract("")
}

