% Generated by roxygen2 (4.0.1): do not edit by hand
\name{debate.txt.count.matches}
\alias{debate.txt.count.matches}
\title{Count all occurences of keywords or sentences within a character string}
\usage{
debate.txt.count.matches(txt, items)
}
\arguments{
\item{txt}{the character string to mine.}

\item{items}{a list of character strings for which we want to count occurences.}
}
\description{
This functions is basically a wrapper for the \code{\link[base]{gregexpr}} function, allowing to count all occurences of several items in one single step.
}
\examples{
path.to.pdf <- paste0(path.package("debate"), "/doc/BOACG_Tome_21_article_49.pdf")
txt <- debate.pdftotext(path.to.pdf)
txt <- debate.txt.clean(txt)
debate.txt.count.matches(txt = txt, items = "président")
debate.txt.count.matches(txt = txt, items = "Président")
debate.txt.count.matches(txt = txt, items = "soumis au vote de la séance plénière de l'Assemblée constituante")
debate.txt.count.matches(txt = txt, items = "qsdfbnmsdfui")
debate.txt.count.matches(txt = txt, items = c("qsdfbnmsdfui", "président", "merci", "soumis au vote"))
}

