% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/txt.tools.r
\name{debate.txt.extract.markup}
\alias{debate.txt.extract.markup}
\title{Extract text contains between two tags}
\usage{
debate.txt.extract.markup(txt, pattern.start, pattern.stop, quiet = TRUE)
}
\arguments{
\item{txt}{the text to mine.}

\item{pattern.start}{the opening tag. Can be a static character string or a regex.}

\item{pattern.stop}{the closing tag. Can be a static character string or a regex.}

\item{quiet}{if \code{FALSE} the function prints for each match both the extracted text and the remaining text.}
}
\description{
This function allows to extract all accurences of text contained between two tags.
}
\examples{
txt <- "bla blo1 blu ble bla blo2 blu ble bla blo3 blu ble bla blo4 blu ble"
pattern.start = "bla"
pattern.stop = "blu"

debate.txt.extract.markup(
  txt,
  pattern.start,
  pattern.stop
)

txt <- "qdsfbsqdfiusd  \\nM. Genecand. Contrib1. \\n\\nLe président. kjdsfpiou,kjsdfp. qsdfkjhi. \\n\\nM. Genecand. Contrib2. \\n\\nLe président. kjdsfpiou,kjsdfp. qsdfkjhi."
pattern.start = "\\nM. Genecand."
pattern.stop = "\\nLe président."

debate.txt.extract.markup(
txt,
pattern.start,
pattern.stop
)
}

