% Generated by roxygen2 (4.0.1): do not edit by hand
\name{debate.txt.clean}
\alias{debate.txt.clean}
\title{Clean a character string by remplacing non-standard characters}
\usage{
debate.txt.clean(txt)
}
\arguments{
\item{txt}{the character string to be clean}
}
\description{
This function replace non-standard characters within a character strings. This especially concerns characters coming from a PDF to TXT conversion.  For example `’` will be replace by `'`.
}
\examples{
debate.txt.clean("--  ’  --  \\\\f  --  -\\\\n  --")
}

